simplify interface to back end
