+++ src/posix.mak
@@ -112,7 +112,7 @@ DMD_OBJS = \
 	man.o arrayop.o port.o response.o async.o json.o speller.o aav.o unittests.o \
 	imphint.o argtypes.o ti_pvoid.o apply.o sideeffect.o \
 	intrange.o canthrow.o \
-	pdata.o cv8.o \
+	pdata.o cv8.o backconfig.o \
 	$(TARGET_OBJS)
 
 ifeq (OSX,$(TARGET))
@@ -159,7 +159,7 @@ SRC = win32.mak posix.mak \
 	$C/dwarf.c $C/dwarf.h $C/aa.h $C/aa.c $C/tinfo.h $C/ti_achar.c \
 	$C/ti_pvoid.c $C/platform_stub.c $C/code_x86.h $C/code_stub.h \
 	$C/machobj.c $C/mscoffobj.c \
-	$C/xmm.h $C/obj.h $C/pdata.c $C/cv8.c \
+	$C/xmm.h $C/obj.h $C/pdata.c $C/cv8.c $C/backconfig.c \
 	$C/md5.c $C/md5.h \
 	$(TK)/filespec.h $(TK)/mem.h $(TK)/list.h $(TK)/vec.h \
 	$(TK)/filespec.c $(TK)/mem.c $(TK)/vec.c $(TK)/list.c \
@@ -245,6 +245,9 @@ async.o: $(ROOT)/async.c
 
 attrib.o: attrib.c
 	$(CC) -c $(CFLAGS) $<
+
+backconfig.o: $C/backconfig.c
+	$(CC) -c $(MFLAGS) $<
 
 bcomplex.o: $C/bcomplex.c
 	$(CC) -c $(MFLAGS) $<
